+++ p2p/server_test.go
@@ -36,20 +36,20 @@ import (
 
 type testTransport struct {
 	rpub *ecdsa.PublicKey
-	*rlpx
+	*frameRW
 
 	closeErr error
 }
 
 func newTestTransport(rpub *ecdsa.PublicKey, fd net.Conn) transport {
-	wrapped := newRLPX(fd).(*rlpx)
+	wrapped := newRLPX(fd).(*frameRW)
 	wrapped.rw = newRLPXFrameRW(fd, secrets{
 		MAC:        zero16,
 		AES:        zero16,
 		IngressMAC: sha3.NewLegacyKeccak256(),
 		EgressMAC:  sha3.NewLegacyKeccak256(),
 	})
-	return &testTransport{rpub: rpub, rlpx: wrapped}
+	return &testTransport{rpub: rpub, frameRW: wrapped}
 }
 
 func (c *testTransport) doEncHandshake(prv *ecdsa.PrivateKey, dialDest *ecdsa.PublicKey) (*ecdsa.PublicKey, error) {
@@ -62,7 +62,7 @@ func (c *testTransport) doProtoHandshake(our *protoHandshake) (*protoHandshake, 
 }
 
 func (c *testTransport) close(err error) {
-	c.rlpx.fd.Close()
+	c.frameRW.fd.Close()
 	c.closeErr = err
 }
